requirements.txt
streamlit
google-genai
deep-translator
gTTS
streamlit-webrtc
SpeechRecognition
sounddevice
numpy

